[Decomp][Player] Cohesion split: extract instance-bind methods to PlayerInstance.cpp
